requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
pandas==2.2.1
openpyxl==3.1.2
pydantic==2.8.2
rapidfuzz==3.6.2
python-multipart==0.0.9
loguru==0.7.2
langchain-google-genai
python-dotenv
ezdxf>=1.3.0
pdfplumber>=0.10.0
python-docx>=1.1.0
langgraph>=0.1.0
langchain-core>=0.2.0
langchain-community>=0.2.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
psycopg2-binary>=2.9.9
pymupdf>=1.23.0
Pillow>=10.0.0

